requirements.txt
paho-mqtt==1.6.1
gitpython==3.1.27
pyyaml==6.0